RAGProject/requirements.txt
## Requirements.txt

fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.23
alembic==1.12.1
pinecone-client==2.2.4
langchain==0.0.340
langchain-community==0.0.1
sentence-transformers==2.2.2
transformers==4.35.2
# torch==2.1.1
torch
groq==0.4.1
python-dotenv==1.0.0
loguru==0.7.2
pydantic==2.5.0
httpx==0.25.2